% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survdat.example}
\alias{survdat.example}
\title{An example sample data file of 5000 respondents.}
\format{An object of class \code{data.frame} with 500 rows and 6 columns.}
\usage{
survdat.example
}
\description{
A dataset containing the values of y (a variable of interest which is outlier-prone),
two auxiliary variables x1 and x2,
piwt (inverse probability of selection weight), and
gregwt (generalized regression estimation weight, assumed to have been
 calculated using x1 and x2)
}
\keyword{datasets}

