% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survdat.example}
\alias{survdat.example}
\title{An example sample data file of 5000 respondents.}
\format{\preformatted{'data.frame':	500 obs. of  5 variables:
 $ y     : num  6.66 40.02 3.47 5.23 1.03 ...
 $ x1    : num  0.689 1.237 1.182 1.336 0.734 ...
 $ x2    : num  0.778 1.961 1.044 1.085 1.218 ...
 $ piwt  : num  2.45 1.81 1.85 1.75 2.36 ...
 $ gregwt: num  2.45 1.81 1.85 1.75 2.36 ...
}}
\usage{
survdat.example
}
\description{
A dataset containing the values of y (a variable of interest which is outlier-prone), and two auxiliary variables x1 and x2.
}
\keyword{datasets}

